Make sure the last multiplication of norms are explicitly using double. The current code is quite unclear, so it is better to be explicit (Math.sqrt uses double argument anyways)

+++ lucene/core/src/java20/org/apache/lucene/util/VectorUtilPanamaProvider.java
@@ -217,7 +217,7 @@ final class VectorUtilPanamaProvider implements VectorUtilProvider {
       norm1 += elem1 * elem1;
       norm2 += elem2 * elem2;
     }
-    return (float) (sum / Math.sqrt(norm1 * norm2));
+    return (float) (sum / Math.sqrt((double) norm1 * (double) norm2));
   }
 
   @Override
